refactor(bosun): the hull contract stops sharing a file with passt argv
